@@ -1,8 +1,8 @@
 ext.versions = [
         minSdk        : 19,
-        targetSdk     : 26,
-        compileSdk    : 26,
-        buildTools    : '26.0.1',
+        targetSdk     : 28,
+        compileSdk    : 28,
+        buildTools    : '28.0.3',
 
         // Google
         supportLibrary: '27.0.2',
@@ -12,15 +12,15 @@ ext.versions = [
         rxAndroid     : '2.0.1',
 ]
 
-ext.bintrayVersion = '2.0.23'
+ext.bintrayVersion = '1.8.4'
 
 ext.libraries = [
 
         // Google
-        supportV7      : "com.android.support:appcompat-v7:$versions.supportLibrary",
-        supportFragment: "com.android.support:support-fragment:$versions.supportLibrary",
-        design         : "com.android.support:design:$versions.supportLibrary",
-        recycler       : "com.android.support:recyclerview-v7:$versions.supportLibrary",
+        supportV7      : 'androidx.appcompat:appcompat:1.0.0',
+        supportFragment: 'androidx.fragment:fragment:1.0.0',
+        design         : 'com.google.android.material:material:1.0.0',
+        recycler       : 'androidx.recyclerview:recyclerview:1.0.0',
 
         // Rx
         rxJava         : "io.reactivex.rxjava2:rxjava:$versions.rxJava",
