@@ -1,12 +1,9 @@
 package com.artemkopan.mvp.presentation;
 
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
 import android.view.View;
-
-import com.artemkopan.mvp.presenter.BasePresenter;
-import com.artemkopan.mvp.view.BaseView;
 
 import io.reactivex.disposables.CompositeDisposable;
 
@@ -21,7 +18,7 @@ public interface Presentation {
     /**
      * if int == 0 then wasn't load inflate
      *
-     * @return {@link android.support.annotation.LayoutRes} layout res id
+     * @return {@link androidx.annotation.LayoutRes} layout res id
      */
     int onInflateLayout();
 
