@@ -1,9 +1,9 @@
 package com.artemkopan.recycler.adapter;
 
-import android.support.annotation.Nullable;
-import android.support.v7.util.SortedList;
-import android.support.v7.util.SortedList.Callback;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.SortedList;
+import androidx.recyclerview.widget.SortedList.Callback;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.Collection;
 
