@@ -2,7 +2,7 @@ package com.artemkopan.recycler.decoration;
 
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class GridSpacesItemDecoration extends RecyclerView.ItemDecoration {
