@@ -12,9 +12,9 @@ import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.graphics.ColorFilter;
 import android.os.Build.VERSION_CODES;
-import android.support.annotation.AnimRes;
-import android.support.annotation.RequiresApi;
-import android.support.v4.util.ArrayMap;
+import androidx.annotation.AnimRes;
+import androidx.annotation.RequiresApi;
+import androidx.collection.ArrayMap;
 import android.util.Property;
 import android.view.View;
 import android.view.ViewGroup;
