@@ -10,9 +10,9 @@ import android.net.NetworkInfo;
 import android.os.Build;
 import android.os.Looper;
 import android.os.PowerManager;
-import android.support.annotation.FloatRange;
-import android.support.v4.text.TextUtilsCompat;
-import android.support.v4.view.ViewCompat;
+import androidx.annotation.FloatRange;
+import androidx.core.text.TextUtilsCompat;
+import androidx.core.view.ViewCompat;
 import android.view.View;
 import android.view.WindowManager;
 import android.view.inputmethod.InputMethodManager;
@@ -21,7 +21,6 @@ import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
-import java.util.Locale;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocketFactory;
