@@ -1,6 +1,6 @@
 package com.artemkopan.utils;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 /**
  * Created by Artem Kopan for jabrool
