@@ -5,9 +5,9 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Parcelable;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.app.Fragment;
-import android.support.v4.util.Pair;
+import androidx.core.app.ActivityCompat;
+import androidx.fragment.app.Fragment;
+import androidx.core.util.Pair;
 import android.view.View;
 
 import com.artemkopan.utils.transitions.TransitionHelper;
@@ -15,7 +15,7 @@ import com.artemkopan.utils.transitions.TransitionHelper;
 import java.io.Serializable;
 import java.util.ArrayList;
 
-import static android.support.v4.app.ActivityCompat.startActivityForResult;
+import static androidx.core.app.ActivityCompat.startActivityForResult;
 
 /**
  * Created by Artem Kopan for BaseProject
