@@ -1,7 +1,7 @@
 package com.artemkopan.utils.view_pager.transforms;
 
 
-import android.support.v4.view.ViewPager.PageTransformer;
+import androidx.viewpager.widget.ViewPager.PageTransformer;
 import android.view.View;
 
 public abstract class ABaseTransformer implements PageTransformer {
