@@ -2,7 +2,7 @@ package com.artemkopan.widget.aspect;
 
 import android.content.Context;
 import android.graphics.Point;
-import android.support.v7.widget.AppCompatImageView;
+import androidx.appcompat.widget.AppCompatImageView;
 import android.util.AttributeSet;
 
 /**
