@@ -2,8 +2,8 @@ package com.artemkopan.widget.places;
 
 import android.content.Context;
 import android.location.Address;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.NonNull;
+import androidx.annotation.LayoutRes;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
